refactor(routes): skills drafts workflow sub-router 분리
# 동기
skills.routes.ts (lint 기준 502 lines) max-lines warning 해소. drafts
workflow 3 endpoints (GET /drafts, POST /:id/approve|reject) 가 가장
self-contained — sub-router 패턴으로 추출.

# 변경
- routes/skills-drafts.routes.ts (신규, 133 lines)
  - GET /drafts (target=user/system/all)
  - POST /:skillId/approve (draft → active)
  - POST /:skillId/reject (draft → archived)
  - extractUserId() helper (PublicUser/AuthUser 양쪽 호환)
- routes/skills.routes.ts
  - 3 endpoint 본문 제거 (-101 lines)
  - skillsDraftsRouter mount: `router.use('/', skillsDraftsRouter)`
  - draftsQuerySchema import 제거 (sub-router 직접 import)
- mount path: 외부에서 보는 URL 동일 (/api/agents/skills/drafts 등)

# 결과
- skills.routes.ts (lint 기준): 502 → 423 lines (-79)
- max 400 너머 23 lines 잔존 — 후속 split 가능 (create flow / user-assign)
- 외부 호출 영향 0 — sub-router mount path 동일

# 검증
- tsc 0 / lint 0 errors / 1 warning (max-lines 423 잔존)
- 전체 회귀: 97 suites / 1604 tests / 0 failed

# 후속 (별도 PR)
- skills-create.routes.ts (auto-create / import-from-git / upload, 4 endpoints)
- skills-assign.routes.ts (user-assigned, 3 endpoints)
- agent-loop-strategy.ts / sockets/handler.ts / request-handler.ts /
  unified-database.ts 의 max-lines 후속 split
